Calculate all rows on the configuration list and support sorting

+++ model/jvm/src/main/scala/explore/modes/SpectroscopyModesMatrixPlatform.scala
@@ -21,7 +21,7 @@ trait SpectroscopyModesMatrixPlatform extends SpectroscopyModesMatrixDecoders {
       .flatMap(l => Stream(l.toList: _*))
       .zipWithIndex
       .map { case (r, i) =>
-        r.copy(id = i)
+        r.copy(id = i.toInt)
       }
       .compile
       .toList
